data_feeders/requirements.txt
websocket-client==0.57.0
pymongo==3.11.3